+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 0.0.116 (2020-05-19 / d75a7cc)
 
 ## Fixed